==> 1-ML-course-project-data-cleaning.ipynb <==
# in this notebook, we are going explore the data
# Data link
# https://www.kaggle.com/petersunga/google-amazon-facebook-employee-reviews/home
import numpy as np
import matplotlib.pyplot as plt
import pandas as pd


df = pd.read_csv(r'C:\Users\Adarsh\uOttawa\ML Course Project-7Apr\datasets\employee_reviews.csv')
                  #sep=',' , engine = 'python')

df.shape

df.head(3)

 # printing all the features
df.columns

# 1 )  keep relevant feature and remove 'Unnamed' and 'link'
df = df[['company', 'location',  'dates', 'job-title', 'summary',
       'pros', 'cons', 'advice-to-mgmt', 'overall-ratings',
       'work-balance-stars', 'culture-values-stars',
       'carrer-opportunities-stars', 'comp-benefit-stars',
       'senior-mangemnet-stars', 'helpful-count']]

# Check the null values
df.isnull().sum().tolist()

# we don't need to remove the null values because those are present only in the text data, after merging all the comments in one
# these null values would be gone

# 2) let's merge all the comments on the companies into one and name it feedback

# there are some columns which are not string first change them intro a string or replace Nan with - and we can  filter - later.
#df['summary'] = df['summary'].map({'Nan': '-'  })
#df['pros'] = df['pros'].map({'Nan': '-'   })
#df['cons'] = df['cons'].map({'Nan': '-'   })
#df['advice-to-mgmt'] = df['advice-to-mgmt'].map({'Nan': '-'   })

df['summary'] = df['summary'].astype(str)
df['pros'] = df['pros'].astype(str)
df['cons'] = df['cons'].astype(str)
df['advice-to-mgmt'] = df['advice-to-mgmt'].astype(str)



#axis : {0 or ‘index’, 1 or ‘columns’}, default 0
#Axis along which the function is applied:
#0 or ‘index’: apply function to each column.
#1 or ‘columns’: apply function to each row.

df['feedback'] = df[['summary' , 'pros' , 'cons' , 'advice-to-mgmt' ]].apply( lambda x : ' '.join( x  ), axis = 1 )



df['feedback'][10]

df.columns

# we can get rid of the features which are included in the feedback
df = df[['company', 'location' , 'dates', 'job-title', 'overall-ratings', 'work-balance-stars',
       'culture-values-stars', 'carrer-opportunities-stars',
       'comp-benefit-stars', 'senior-mangemnet-stars', 'helpful-count',
       'feedback']]

df.shape

df.head(5)

# Feature Engineering, 
# extract features from existing features

df['job-title'].value_counts()

# let's make a new feature job_status which is  going to store if the employe is current or former eomployee

job_status = []
designation = []
des_string = ''

for i in range(len(df['job-title'])):
    if "Current Employee" in df['job-title'][i]:
        job_status.append("Current Employee")
        des_string = df['job-title'][i][18:]
        designation.append(des_string)
        
    elif "Former Employee" in df['job-title'][i]:
        job_status.append("Former Employee")
        des_string = df['job-title'][i][17:]
        designation.append(des_string)
        
    
#print(job_status)

# merge the new feature to the datafram
df['job_status'] = job_status
df['job_title'] = designation
df.head()

# convert the 'dates' feature to pandas date time format and then extract day, month and year as features
#df['dates'] = pd.to_datetime(df.dates , format='%b %d, %Y')


# make 3 new features, day_of_the_month, month, year and extract them from the dates feature whose values are in the format
# of string

post_day = []
post_month = []
post_year = []
temp = ''

for i in range(len(df['dates'])):
    temp = df['dates'][i][1:4]
    post_month.append(temp)

    temp = df['dates'][i][-8:-6]
    post_day.append(temp)    
    
    temp = df['dates'][i][-5:]
    post_year.append(temp)
       

# merge the new feature to the datafram
df['post_day'] = post_day
df['post_month'] = post_month
df['post_year'] = post_year
df.head()

df['job_status'].value_counts()

# if we are considering job_status as our Target for classification then surely it;s a class impabalance problem
# let's do undersampling to keep all the values equal for both the classes 
# chose 20000 values from both the classes randomly

# use resample method from scikit-learn
from sklearn.utils import resample

# Separate majority and minority classes for undersampling
df_Current_Employee  = df[df.job_status== 'Current Employee']
df_Former_Employee   = df[df.job_status == 'Former Employee']
    
df_Current_Employee = resample(df_Current_Employee, 
                                             replace=True,    # sample with replacement
                                             n_samples=20000,     # to match number of values in each class
                                             random_state=123) # reproducible results

df_Former_Employee   = resample(df_Former_Employee, 
                                             replace=True,    # sample with replacement
                                             n_samples=20000,     # to match number of values in each class
                                             random_state=123) # reproducible results

# Combine all the class with equal number of values
df_sampled = pd.concat([ df_Current_Employee , df_Former_Employee])
 

# check for the class balance
print(df_sampled['job_status'].value_counts())
# change df with sampled one
df= df_sampled
print("\n")
print(df.head(5))

df.shape

df.post_month.value_counts()

# Map months in string to corresponding number in order
df['post_month'] = df['post_month'].map({'Jan': 1 ,
                                         'Feb': 2 ,
                                         'Mar' : 3,
                                         'Apr' : 4,
                                         'May' : 5,
                                         'Jun' : 6,
                                         'Jul' : 7,
                                         'Aug' : 8,
                                         'Sep' : 9,
                                         'Oct' : 10,
                                         'Nov' : 11,
                                         'Dec' : 12,
                                        'None' : 'None'})


df.head(5)

df.columns

# remove the columns which are merged, or new columns created from them

df = df[['company', 'location', 'overall-ratings',
       'work-balance-stars', 'culture-values-stars',
       'carrer-opportunities-stars', 'comp-benefit-stars',
       'senior-mangemnet-stars', 'helpful-count', 'feedback', 'job_status',
       'job_title', 'post_day', 'post_month', 'post_year' ]]

df.shape

# work on the "none" values in the rows
df['location'] = df['location'].replace(['none'], 'Unknown')

df['work-balance-stars'].value_counts()

# as all the ratings have "none" as a value except overall-ratings, we need to replace it with the average rating value 2.5
# reason: 2.5 is not too high not too low, not influence the models in a significant way, that's what we want here
df['work-balance-stars'] = df['work-balance-stars'].replace(['none'], 2.5)
df['culture-values-stars'] = df['culture-values-stars'].replace(['none'], 2.5)
df['carrer-opportunities-stars'] = df['carrer-opportunities-stars'].replace(['none'], 2.5)
df['comp-benefit-stars'] = df['comp-benefit-stars'].replace(['none'], 2.5)
df['senior-mangemnet-stars'] = df['senior-mangemnet-stars'].replace(['none'], 2.5)

# work on the none/blank/error values of the dates (day, month, year)
df.post_day.value_counts().sort_values()

# replace the day of the month 0 and blank with 15
df['post_day'] = df['post_day'].replace([' 0' , ''], 15)

# check if the values been replaced
#change the type of the post_day from string to int
df.post_day = df.post_day.astype(np.int64)

df.post_day.value_counts().sort_values()


# same process as post_day goes for post month and year
print(df.post_month.value_counts())
print("\n")
print(df.post_year.value_counts())


# check for null values in post month
# Check the null values
df.isnull().sum().tolist()

# Aaah! there is one null value in the month column

# Drop null values
df = df.dropna(how="any")

# as there are no none values for month, we just need to work on the year
# replace 0000, None with 2009, as there are only 4 such values, so it doesnt matter much where to put the values

df['post_year'] = df['post_year'].replace([' 0000' , 'None'], '2009')


# meanwhile, also change the type from str to int for both month and year
df.post_month = df.post_month.astype(np.int64)
df.post_year = df.post_year.astype(np.int64)

print(df.post_month.value_counts())
print("\n")
print(df.post_year.value_counts())


#### "String to Numeric"  -> Convert the features which are supposed to be numeric but are string
# let's convert the overall ratings which are in string to numeric

pd.to_numeric(pd.Series((df['overall-ratings'])), errors='coerce')
pd.to_numeric(  pd.Series((df['work-balance-stars'])) , errors='coerce')
pd.to_numeric(  pd.Series((df['culture-values-stars'])) , errors='coerce')
pd.to_numeric(  pd.Series((df['carrer-opportunities-stars'])) , errors='coerce')
pd.to_numeric(  pd.Series((df['comp-benefit-stars'])) , errors='coerce')
pd.to_numeric(  pd.Series((df['senior-mangemnet-stars'])) , errors='coerce')

# change the help count to numeric as well
pd.to_numeric(  pd.Series((df['helpful-count'])) , errors='coerce')

df.head(5)

df.to_csv(r'C:\Users\Adarsh\uOttawa\ML Course Project-7Apr\datasets\1-cleaned_employee_reviews.csv' , index=False)
