==> 4a-text-data-exploration-processing.ipynb <==
# in this file we are goona work on the text data, extract features from text and build models

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

from sklearn.feature_extraction.text import CountVectorizer

from textblob import TextBlob

df = pd.read_csv(r'C:\Users\Adarsh\uOttawa\ML Course Project-7Apr\datasets\2_text_data.csv')

df.shape

df.columns

df.tail(20)

# 1. Feature extraction : number of words
# basic idea is to extract number words in each title 
# intuition behind with this technique is - some collection/categories need longer title than others

# make a new feature  word_count and calculate the words by splitting the strin on white space
df['feedback_word_count'] = df['feedback'].apply( lambda x : len(str(x).split(" ") ) )

df.head(5)

# 2. Feature : Average word length (Sum categories might need longer words than others)
# formula : Sum (length of all the words in the tweet or doc ) / (total length of the tweet or doc)

def avg_word_length(sentence):
    words = sentence.split()
    return ( sum( len(word) for word in words ) / len(words))

df['avg_word_len'] = df['feedback'].apply( lambda x: avg_word_length(x) )

# print
df.head(5)

# change type of feedback from object to string to avoid future errors
df['feedback'] = df['feedback'].astype('str') 

def sentiment(x):
    sentiment = TextBlob(x)
    return sentiment.sentiment.polarity

df['sentiment_score'] = df['feedback'].apply(sentiment)


df.head(5)

# assign numeric values to the sentiments

sentiment_numeric = []

for score in df['sentiment_score']:
    # positive sentiments will have value "+1"
    if score > 0:
        sentiment_numeric.append(1)
    # positive sentiments will have value "-1"
    elif score < 0:
        sentiment_numeric.append(-1)
    # positive sentiments will have value "0"
    elif score == 0:
        sentiment_numeric.append(0)

df['sentiment'] = sentiment_numeric
df.sentiment.value_counts()

df.head(5)

# let's start cleaning the data 

# 1. Lower casing  - change all the words to lower case to avoid duplication. Because "Python" and "python" considered 2 words
# we will split the title into words and then convert those words into lower case and then join

df['feedback'] = df['feedback'].apply( lambda x : " ".join( x.lower() for x in x.split() ) )
df['feedback'].head(5)

# 2. Remove punctuations and special charactaers
df['feedback'] = df['feedback'].str.replace('[^\w\s]','')
# the [^\w\s] means remove everything, keep only words(w) and spaces(s)
# this step should be done after feature extraction like hashtags, user tagged
df['feedback'].head()

# 3 . Removal of stop words - the, a , and etc. These are most commonly occuring words, and may created irrevelent 
# baises to our model

#import nltk
# nltk.download('stopwords')
from nltk.corpus import stopwords
stop = stopwords.words('english')

# we are gonna split the title into words and then filter the stop words and join them back
df['feedback'] = df['feedback'].apply( lambda x : " ".join( x for x in x.split() if x not in stop) )
df['feedback'].head(5)

# 4. Frequent word removal from the text; text which are not stopwords
# first we will take whole tweets data and split into words and then calculate their frequency
#  join words only with strings so, there needs to be some string to join other string
all_words = ' '.join( df['feedback'] ).split()
freq = pd.Series(all_words).value_counts()[:10]
freq

# 4a)the freq words might add some bias to our model and not contribute to our model 
# remove those common words which are  ouccring more than 10,000 times using "freq"
df['feedback'] = df['feedback'].apply( lambda x: " ".join( x for x in x.split() if x not in freq ) )
df.head(5)

# 4c) remove none word, becuase it was given when title was not present properly
df['feedback'] = df['feedback'].apply( lambda x: " ".join( x for x in x.split() if x not in ["none", "None"] ) )
df.head(5)


# 5. Remove Rare words, which will not contribute to our model
all_words = ' '.join( df['feedback'] ).split()
rarely = pd.Series(all_words).value_counts()[-55000:]
rarely.sort_values

df['feedback'].head(5)

# the abover words might not contribute to our model
# remove those rare words 
df['feedback1'] = df['feedback'].apply( lambda x: " ".join( x for x in x.split() if x not in rarely ) )


df['feedback1'].head(15)

df.feedback.head(5)

# Check again for  rare words
all_words = ' '.join( df['feedback'] ).split()
rarely = pd.Series(all_words).value_counts()
rarely.sort_values

# 6. Spelling correction  -- not  necessary just need to remove the typos (rare words, which are accidently introduced)

# most common issue to deal with; spelling mistake, typos, shortcuts, abbreviatiosn (very common in when typing)
# spelling correction also help remove duplication that created by spelling mistakes like python, pythn, pythom etc.

# ** TextBlob is a Python library for processing textual data.
# Can be used for common NLP tasks such as part-of-speech tagging, noun phrase extraction, sentiment analysis, classification,
# translation, and more.

# Spelling correction takes a lot of time. Also, we cannot always expect it to be accurate. 
# words are often used in their abbreviated form. For instance, ‘your’ is used as ‘ur’. We should treat this before the spelling
# correction step, otherwise these words might be transformed into any other word.

from textblob import TextBlob

# let's try spelling correction on 5 titles
# TextBlob takes one title and checks for each word for correction and corrects spellings and returns TextBlob; we can convert 
# into a string
#type(TextBlob("ysfunctional selfish drags kiss dysfu").correct())


# !!!!!!!!!!!!!!!!!!!!!!!!  Warning: This step is gonna take too much time maybe a few hours
#train['title'].apply( lambda x :  str( TextBlob(x).correct() ) )


# 7 . Lemmatization ; its preferred over stemming because if finds the root word
import nltk
nltk.download('wordnet')
from textblob import Word 



# perform lemmatization on the title
df['feedback'] = df['feedback'].apply( lambda x : " ".join( [Word(word).lemmatize() for word in x.split()  ]) )
df['feedback'].head(5)

df.head(5)

# better save the file to avoid time consuming re-work on lemmatization
df.to_csv(r'C:\Users\Adarsh\uOttawa\ML Course Project-7Apr\datasets\4a-after-text-cleaning.csv', 
                    index=False)

# Now our data cleaning has almost been done. It's time to extract more features :
# 1. n-grams
# 2. tf-idf
# 3. bag of words

df = pd.read_csv(r'C:\Users\Adarsh\uOttawa\ML Course Project-7Apr\datasets\4a-after-text-cleaning.csv')

df.head(5)

# calculating tf-idf using scikitlearn
# sublinear_df  - is set to True to use a logarithmic form for frequency.
# min_df - is the minimum numbers of documents a word must be present in to be kept.
# norm - is set to l2, to ensure all our feature vectors have a euclidian norm of 1.
# ngram_range -  is set to (1, 2) to indicate that we want to consider both unigrams and bigrams.
# stop_words - is set to "english" to remove all common pronouns ("a", "the", ...) to reduce the number of noisy features.

from sklearn.feature_extraction.text import TfidfVectorizer

tfidf = TfidfVectorizer(sublinear_tf=True, min_df=50, norm='l2', encoding='latin-1', ngram_range=(1, 3))

features = tfidf.fit_transform(df.feedback.values.astype('str')).toarray()
labels = df.job_status


# create a dataframe to lookup which label value belongs to which collection?
#category_id_df = df[['Product', 'category_id']].drop_duplicates().sort_values('category_id')
job_status_df = df[['job_status']].drop_duplicates().sort_values('job_status')


category_ids = job_status_df.values
#id_to_category = dict(collection_labels_df[['collection', 'collection_labels']].values)

category_ids


# let's find some interesting corelated unigrams and bigrams and trigrams
from sklearn.feature_selection import chi2
import numpy as np

N = 5

for status in sorted(category_ids):
    
    
    features_chi2 = chi2(features, labels == labels)
    indices = np.argsort(features_chi2[0])
    feature_names = np.array(tfidf.get_feature_names())[indices]
    unigrams = [v for v in feature_names if len(v.split(' ')) == 1]
    bigrams = [v for v in feature_names if len(v.split(' ')) == 2]
    trigrams = [v for v in feature_names if len(v.split(' ')) == 3]
    
    print("# '{}':".format(status))
    print("  . Most correlated unigrams:\n       . {}".format('\n       . '.join(unigrams[-N:])))
    print("  . Most correlated bigrams:\n       . {}".format('\n       . '.join(bigrams[-N:])))
    print("  . Most correlated trigrams:\n       . {}".format('\n       . '.join(trigrams[-N:])))

# further data exploration
cv = CountVectorizer(stop_words='english')
data_cv = cv.fit_transform(df.feedback.values.astype('str'))

data_dtm = pd.DataFrame(data_cv.toarray(), columns=cv.get_feature_names())
data_dtm.index = df.index


data_dtm.columns
