==> 4b-text-data-predictive-modelling.ipynb <==
# in this file we are goona work on the text data, extract features from text and build models
# Now our data cleaning has almost been done. It's time to extract more features :
# 1. n-grams
# 2. tf-idf
# 3. bag of words


import pandas as pd
import numpy as np
import matplotlib.pyplot as plt

# let's do some moedelling
from sklearn.model_selection import train_test_split, cross_val_score

# packages for metric for evalaution of the models
from sklearn import metrics
from sklearn.metrics import classification_report , accuracy_score, confusion_matrix


from sklearn.feature_extraction.text import CountVectorizer
from sklearn.feature_extraction.text import TfidfTransformer

# As tf–idf is very often used for text features, there is also another class called TfidfVectorizer that combines 
# all the options of CountVectorizer and TfidfTransformer in a single model.
# so we can calculate tf-idf and the CountVectorizer in one go with below module
from sklearn.feature_extraction.text import TfidfVectorizer

# use multiple models
from sklearn.linear_model import LogisticRegression
from sklearn.ensemble import RandomForestClassifier
from sklearn.naive_bayes import MultinomialNB
from sklearn.svm import LinearSVC, SVC

import seaborn as sns

# to save the models so that we don't need to train the models every time we need to do prediction
from sklearn.externals import joblib




df = pd.read_csv(r'C:\Users\Adarsh\uOttawa\ML Course Project-7Apr\datasets\4a-after-text-cleaning.csv')

df.head(5)

# calculating tf-idf using scikitlearn
# sublinear_df  - is set to True to use a logarithmic form for frequency.
# min_df - is the minimum numbers of documents a word must be present in to be kept.
# norm - is set to l2, to ensure all our feature vectors have a euclidian norm of 1.
# ngram_range -  is set to (1, 2) to indicate that we want to consider both unigrams and bigrams.
# stop_words - is set to "english" to remove all common pronouns ("a", "the", ...) to reduce the number of noisy features.


tfidf = TfidfVectorizer(sublinear_tf=True, min_df=50, norm='l2', encoding='latin-1', ngram_range=(1, 3))

features = tfidf.fit_transform(df.feedback.values.astype('str')).toarray()
labels = df.job_status

features.shape

features

# convert features ndarray into pandas dataframe so that it can be merged with 
#other features like avg_word_count and avg_word_len
features_df = pd.DataFrame(features)

features_df.tail(5)

df_word = df[['feedback_word_count' , 'avg_word_len']]

X_train, X_test, y_train, y_test = train_test_split(df_word, labels, random_state = 0)

# let's try how ensemling ( Random forest wit 10 decision trees ) performs 
random_forest = RandomForestClassifier(n_estimators=10)
random_forest.fit(X_train, y_train)

print("Accuracy of the random forest with 10 DTs, with only 2 features considered - word count and world length")
print("Accuracy = ")
print(random_forest.score(X_test, y_test))
print("\n")
y_pred_rf = random_forest.predict(X_test)
print(classification_report(y_test, y_pred_rf))

df_word.shape

type(df_word)

df_word.head(5)

final_features = pd.concat([features_df, df_word], axis=1)

final_features.shape

X_train, X_test, y_train, y_test = train_test_split(final_features, labels, random_state = 0)

models = [
    RandomForestClassifier(n_estimators=200, max_depth=3, random_state=0),
    LinearSVC(),
    MultinomialNB(),
    LogisticRegression(random_state=0),
]

# define cross validation
CV = 5
cv_df = pd.DataFrame(index=range(CV * len(models)))
entries = []

for model in models:
  model_name = model.__class__.__name__
  accuracies = cross_val_score(model, final_features, labels, scoring='accuracy', cv=CV)
    
  for fold_idx, accuracy in enumerate(accuracies):
    entries.append((model_name, fold_idx, accuracy))

cv_df = pd.DataFrame(entries, columns=['model_name', 'fold_idx', 'accuracy'])
cv_df.groupby('model_name').accuracy.mean()


cv_df

X_train, X_test, y_train, y_test, indices_train, indices_test = train_test_split(final_features, labels, 
                                                                                 df.index, test_size=0.25, random_state=0)


# As LogisticRegression() works better than other let's look deeper
lr = LogisticRegression()
lr.fit(X_train, y_train)

y_pred = lr.predict(X_test)

#conf_mat = confusion_matrix(y_test, y_pred)


print(metrics.classification_report(y_test, y_pred))

# save the logistic regression model
filepath = r'C:\Users\Adarsh\uOttawa\ML Course Project-7Apr\saved_models\text-data\LogisticRegression_model'
joblib.dump(lr ,filepath )
