==> 6b-statistical-testing.ipynb <==
# Chosing statistical test according to our dataset

As we have only one dataset, we need to use test which can test the models on multiple folds of single datasets

Paired t-test would most suitable statistical test for this job


# import libraries

# data manupulation and maths
import pandas as pd
import numpy as np

# import stats for statistics related tasks
from scipy import stats

# plotting
import matplotlib.pyplot as plt
from matplotlib import pyplot
import seaborn as sns

import pprint
# loading and saving the results
from sklearn.externals import joblib


# load 10 fold results of meta-models for statistical testing
filepath = r'C:\Users\Adarsh\uOttawa\ML Course Project-7Apr\intermediate_data\10-fold-accuracies-meta-models'
results = joblib.load(filepath)


results

# Separate results for every model
res_RandomForest  = (results['accuracy'][results.model_name== 'RandomForestClassifier']).values
res_Bagging  = results['accuracy'][results.model_name== 'BaggingClassifier'].values
res_AdaBoost  = results['accuracy'][results.model_name== 'AdaBoostClassifier'].values
res_Voting  = results['accuracy'][results.model_name== 'VotingClassifier'].values




rf_bg = res_RandomForest - res_Bagging
rf_adb = res_RandomForest - res_AdaBoost
rf_vot = res_RandomForest - res_Voting
bg_adb = res_Bagging - res_AdaBoost
bg_vot = res_Bagging - res_Voting
adb_vot = res_AdaBoost - res_Voting

print("     rf-bg         rf-adb         rf-vot       bg-adb         bg-vot        adb-vot")
print("----------------------------------------------------------------------------------")
for i in range(0, len(rf_bg)):
    print("     "+str(round(rf_bg[i],4))+"        "+str(round(rf_adb[i],4))+"        "+str(round(rf_vot[i],4))
                    +"        "+str(round(bg_adb[i],4))+"        "+str(round(bg_vot[i],4))+"        "+str(round(adb_vot[i],4)))

    
    
print("----------------------------------------------------------------------------------")
print("mean = "+str(round(rf_bg.mean(),4))+"        "+str(round(rf_adb.mean(),4))+"        "+str(round(rf_vot.mean(),4))
                    +"        "+str(round(bg_adb.mean(),4))+"        "+str(round(bg_vot.mean(),4))+"        "
      +str(round(adb_vot.mean(),4)))

print("std = "+str(round(rf_bg.std(),4))+"        "+str(round(rf_adb.std(),4))+"        "+str(round(rf_vot.std(),4))
                    +"        "+str(round(bg_adb.std(),4))+"        "+str(round(bg_vot.std(),4))+"        "
      +str(round(adb_vot.std(),4)))




## Cross Checking with the internal scipy function
# Chosing a threshold alpha for our models 
# if p > alpha - > then there's no significance difference between two models
# if p < alpha -> then there is  significance difference between two models

alpha = 0.005
sig_diff = ''

t, p = stats.ttest_ind( res_RandomForest , res_Bagging )
if p > alpha:
    sig_diff = '  No, there is no significant difference'
else:
    sig_diff = '  Yes, there is significant difference'
    
print("(Random Forest - Bagging)   t = " + str(round(t,4)) + "  p = " + str(round(p,4))+sig_diff)

t, p = stats.ttest_ind( res_RandomForest , res_AdaBoost )
if p > alpha:
    sig_diff = '  No, there is no significant difference'
else:
    sig_diff = '  Yes, there is significant difference'
print("(Random Forest - AdaBoost)  t = " + str(round(t,4)) + "  p = " + str(round(p,4))+sig_diff)

t, p = stats.ttest_ind( res_RandomForest , res_Voting )
if p > alpha:
    sig_diff = '  No, there is no significant difference'
else:
    sig_diff = '  Yes, there is significant difference'
print("(Random Forest - Voting)    t = " + str(round(t,4)) + "  p = " + str(round(p,4))+sig_diff)

t, p = stats.ttest_ind( res_Bagging, res_AdaBoost )
if p > alpha:
    sig_diff = '  No, there is no significant difference'
else:
    sig_diff = '  Yes, there is significant difference'
print("(Bagging - AdaBoost)        t = " + str(round(t,4)) + "  p = " + str(round(p,4))+sig_diff)

t, p = stats.ttest_ind( res_Bagging , res_Voting )
if p > alpha:
    sig_diff = '  No, there is no significant difference'
else:
    sig_diff = '  Yes, there is significant difference'
print("(Bagging - Voting)          t = " + str(round(t,4)) + "  p = " + str(round(p,4))+sig_diff)

t, p = stats.ttest_ind( res_AdaBoost , res_Voting )
if p > alpha:
    sig_diff = '  No, there is no significant difference'
else:
    sig_diff = '  Yes, there is significant difference'
print("(AdaBoost - Voting)         t = " + str(round(t,4)) + "  p = " + str(round(p,4))+sig_diff)
